extract updateBlockCommand
