+++ packages/core/src/blocks/ListItemBlockContent/CheckListItemBlockContent/CheckListItemBlockContent.ts
@@ -1,4 +1,5 @@
 import { InputRule } from "@tiptap/core";
+import { updateBlockCommand } from "../../../api/blockManipulation/updateBlock.js";
 import { getCurrentBlockContentType } from "../../../api/getCurrentBlockContentType.js";
 import {
   PropSchema,
@@ -49,12 +50,14 @@ const checkListItemBlockContent = createStronglyTypedTiptapNode({
           }
 
           chain()
-            .BNUpdateBlock(state.selection.from, {
-              type: "checkListItem",
-              props: {
-                checked: false as any,
-              },
-            })
+            .command(
+              updateBlockCommand(this.options.editor, state.selection.from, {
+                type: "checkListItem",
+                props: {
+                  checked: false as any,
+                },
+              })
+            )
             // Removes the characters used to set the list.
             .deleteRange({ from: range.from, to: range.to });
         },
@@ -67,12 +70,14 @@ const checkListItemBlockContent = createStronglyTypedTiptapNode({
           }
 
           chain()
-            .BNUpdateBlock(state.selection.from, {
-              type: "checkListItem",
-              props: {
-                checked: true as any,
-              },
-            })
+            .command(
+              updateBlockCommand(this.options.editor, state.selection.from, {
+                type: "checkListItem",
+                props: {
+                  checked: true as any,
+                },
+              })
+            )
             // Removes the characters used to set the list.
             .deleteRange({ from: range.from, to: range.to });
         },
@@ -82,18 +87,21 @@ const checkListItemBlockContent = createStronglyTypedTiptapNode({
 
   addKeyboardShortcuts() {
     return {
-      Enter: () => handleEnter(this.editor),
+      Enter: () => handleEnter(this.options.editor),
       "Mod-Shift-9": () => {
         if (getCurrentBlockContentType(this.editor) !== "inline*") {
           return true;
         }
 
-        return this.editor.commands.BNUpdateBlock(
-          this.editor.state.selection.anchor,
-          {
-            type: "checkListItem",
-            props: {},
-          }
+        return this.editor.commands.command(
+          updateBlockCommand(
+            this.options.editor,
+            this.editor.state.selection.anchor,
+            {
+              type: "checkListItem",
+              props: {},
+            }
+          )
         );
       },
     };
@@ -212,12 +220,14 @@ const checkListItemBlockContent = createStronglyTypedTiptapNode({
         }
 
         if (typeof getPos !== "boolean") {
-          this.editor.commands.BNUpdateBlock(getPos(), {
-            type: "checkListItem",
-            props: {
-              checked: checkbox.checked as any,
-            },
-          });
+          this.editor.commands.command(
+            updateBlockCommand(this.options.editor, getPos(), {
+              type: "checkListItem",
+              props: {
+                checked: checkbox.checked as any,
+              },
+            })
+          );
         }
       };
       checkbox.addEventListener("change", changeHandler);
