+++ packages/core/src/blocks/ListItemBlockContent/NumberedListItemBlockContent/NumberedListItemBlockContent.ts
@@ -1,4 +1,5 @@
 import { InputRule } from "@tiptap/core";
+import { updateBlockCommand } from "../../../api/blockManipulation/updateBlock.js";
 import { getCurrentBlockContentType } from "../../../api/getCurrentBlockContentType.js";
 import {
   PropSchema,
@@ -44,10 +45,12 @@ const NumberedListItemBlockContent = createStronglyTypedTiptapNode({
           }
 
           chain()
-            .BNUpdateBlock(state.selection.from, {
-              type: "numberedListItem",
-              props: {},
-            })
+            .command(
+              updateBlockCommand(this.options.editor, state.selection.from, {
+                type: "numberedListItem",
+                props: {},
+              })
+            )
             // Removes the "1." characters used to set the list.
             .deleteRange({ from: range.from, to: range.to });
         },
@@ -57,18 +60,21 @@ const NumberedListItemBlockContent = createStronglyTypedTiptapNode({
 
   addKeyboardShortcuts() {
     return {
-      Enter: () => handleEnter(this.editor),
+      Enter: () => handleEnter(this.options.editor),
       "Mod-Shift-7": () => {
         if (getCurrentBlockContentType(this.editor) !== "inline*") {
           return true;
         }
 
-        return this.editor.commands.BNUpdateBlock(
-          this.editor.state.selection.anchor,
-          {
-            type: "numberedListItem",
-            props: {},
-          }
+        return this.editor.commands.command(
+          updateBlockCommand(
+            this.options.editor,
+            this.editor.state.selection.anchor,
+            {
+              type: "numberedListItem",
+              props: {},
+            }
+          )
         );
       },
     };
